Adding a DevOps script `git-revert`
This is one I get asked about quite a bit by other teammates: "How do I roll back a commit".  With this, there is a prompt that asks you everything you need to perform the rollback, and a final confirmation prompt will ask you to make sure you want to do it.

Supports `soft` and `hard` rollbacks, as well as specifying a specific SHA Hash, or using HEAD.  If using HEAD, you can select up to 5 commits to rollback, as that seemed more than enough.

VS Code Command Palette Generates a Dynamic Form to get the required info to generate the command.  Hopefully that's helpful to some folks.